Added the option to run via the command line with the database as an argument.
Added the option to run via the command line. If a user passes in a database file as an argument, we don't create the main window, but instead route all reports to the console. Ctrl+C shuts down the server and exits. Bumped the version number to 0.3.3 for this fairly important minor update.