Added the notion of Spawn/Despawn to GameObjects.
Added the notion of Spawn and Despawn to GameObjects to better manage when their Timers are started and stopped. This cleans up a few other things as well. Player death is handled and communicated better now as well. Plus various bug fixes, text polish, and other tweaks from the first pass at combat.